refactor(`PublicAnnouncement`): Factor out to separate file
