+++ src/models/blockchain/type_scripts/time_lock.rs
@@ -695,7 +695,7 @@ pub mod neptune_arbitrary {
     use super::super::native_currency_amount::NativeCurrencyAmount;
     use super::*;
     use crate::models::blockchain::transaction::transaction_kernel::TransactionKernelModifier;
-    use crate::models::blockchain::transaction::PublicAnnouncement;
+    use crate::models::blockchain::transaction::public_announcement::PublicAnnouncement;
 
     impl Arbitrary for TimeLockWitness {
         /// Parameters are:
