+++ src/models/state/wallet/address/common.rs
@@ -10,7 +10,7 @@ use twenty_first::prelude::Digest;
 
 use crate::config_models::network::Network;
 use crate::models::blockchain::shared::Hash;
-use crate::models::blockchain::transaction::PublicAnnouncement;
+use crate::models::blockchain::transaction::public_announcement::PublicAnnouncement;
 use crate::models::state::wallet::utxo_notification::UtxoNotificationPayload;
 use crate::prelude::twenty_first;
